[MOV] attachment_preview: Reorganize JS modules
