+++ attachment_preview/__manifest__.py
@@ -16,8 +16,11 @@
         "web._assets_primary_variables": [],
         "web.assets_backend": [
             "attachment_preview/static/src/js/attachmentPreviewWidget.esm.js",
-            "attachment_preview/static/src/js/models/attachment_card/attachment_card.esm.js",
-            "attachment_preview/static/src/js/components/chatter/chatter.esm.js",
+            "attachment_preview/static/src/js/utils.esm.js",
+            "attachment_preview/static/src/js/mail_models/attachment_card.esm.js",
+            "attachment_preview/static/src/js/mail_models/attachment_list.esm.js",
+            "attachment_preview/static/src/js/web_views/binary_field.esm.js",
+            "attachment_preview/static/src/js/web_views/form_renderer.esm.js",
             "attachment_preview/static/src/scss/attachment_preview.scss",
             "attachment_preview/static/src/xml/attachment_preview.xml",
         ],
